change the last change

+++ GameEngine/GameEngine/Engine/Render/Pipelining/Render/RenderingPass.cpp
@@ -22,7 +22,7 @@ namespace AGE
 	{
 		_frame_buffer.bind();
 		if (!_is_update) {
-			glDrawBuffers(_drawing_attach.size(), _drawing_attach.data());
+			glDrawBuffers(GLint(_drawing_attach.size()), _drawing_attach.data());
 			for (auto &storage : _frame_output) {
 				_frame_buffer.attachment(*storage.second.get(), storage.first);
 			}
